requirements.txt
flask
matplotlib
pillow
geopy
requests
pyrebase
pyyaml